feat: refactored request docs and skill controller swagger docs
